Resolved the issues and added the mobile number in backend for user

+++ application/controllers/Home.php
@@ -301,14 +301,28 @@ class Home extends CI_Controller
             $this->user_model->update_account_settings($this->session->userdata('user_id'));
             redirect(site_url('home/profile/user_credentials'), 'refresh');
         } elseif ($param1 == "update_photo") {
+
             if (isset($_FILES['user_image']) && $_FILES['user_image']['name'] != "") {
+
+                // Get image file extension
+                $file_extension = pathinfo($_FILES["user_image"]["name"], PATHINFO_EXTENSION);            
+                $allowed_image_extension = array("png","jpg","jpeg");
+
+                if(! in_array($file_extension, $allowed_image_extension)){
+                     $this->session->set_flashdata('error_message', site_phrase('please_enter_valid_image_extension_such_as_.png_.jpg_.jpeg'));
+                     redirect(site_url('home/profile/user_photo'), 'refresh');
+                }
                 unlink('uploads/user_image/' . $this->db->get_where('users', array('id' => $this->session->userdata('user_id')))->row('image') . '.jpg');
                 $data['image'] = md5(rand(10000, 10000000));
                 $this->db->where('id', $this->session->userdata('user_id'));
                 $this->db->update('users', $data);
                 $this->user_model->upload_user_image($data['image']);
+                $this->session->set_flashdata('flash_message', site_phrase('updated_successfully'));
+
+            }else{
+                $this->session->set_flashdata('error_message', site_phrase('please_select_your_profile_image'));
             }
-            $this->session->set_flashdata('flash_message', site_phrase('updated_successfully'));
+
             redirect(site_url('home/profile/user_photo'), 'refresh');
         }
     }
